requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
finance-datareader==0.9.96
pandas
numpy
OpenDartReader
apscheduler==3.10.4